✔️ OK ✔️

+++ C02/ex01/ft_strncpy.c
@@ -12,10 +12,10 @@
 
 char	*ft_strncpy(char *dest, char *src, unsigned int n)
 {
-	unsigned int	i;
+	int	i;
 
 	i = -1;
-	while ((src[++i] != '\0') && (i < n))
+	while ((src[++i] != '\0') && (i < (int)n))
 		dest[i] = src[i];
 	while (i < n)
 		dest[i++] = '\0';
